Added the command CA_GetExtraType, that can be used for numerous things, such as distinguishing signposts from chests. Fixed a bug in CA_GetAgentDanger.
Also added the preparations for a command that will be added soon, GetNearestPlayerNumberToCoords(), which will function much like the other GetNearestXToCoords() functions.

+++ trunk/GW Client API/Main.h
@@ -50,7 +50,7 @@ template <typename T> T ReadPtrChain(
 extern CSectionA* MySectionA;
 extern AgentArray Agents;
 extern AgentMovementArray AgentMovements;
-extern short* AgentTargets;
+extern long* AgentTargets;
 
 #include "MapOverlay.h"
 #include "Item.h"
